requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
argcomplete==3.1.1
async-timeout==4.0.3
attrs==23.1.0
catkin-pkg==0.5.2
certifi==2023.5.7
charset-normalizer==3.2.0
docutils==0.20.1
empy==3.3.4
frozenlist==1.4.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-tools==1.56.0
grpclib==0.4.5
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
lark-parser==0.12.0
lxml==4.9.3
msgpack==1.0.5
multidict==6.0.4
numpy==1.25.0
opencv-contrib-python==4.5.5.62
packaging==23.1
Pillow==10.0.0
ply==3.11
protobuf==4.23.4
pyagxrobots==0.8.4
pycairo==1.24.0
PyGObject==3.44.1
PyJWT==2.8.0
pylimo==0.3.5
pyparsing==3.1.0
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pyserial==3.5
python-can==4.2.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sip==6.7.9
six==1.16.0
TBB==0.2
twilio==8.7.0
typing_extensions==4.7.1
urllib3==2.0.4
viam-sdk==0.4.7
wrapt==1.15.0
yarl==1.9.2
